requirements.txt
torch>=1.12.0
numpy>=1.20.0
scikit-learn>=1.0.0
umap-learn>=0.5.3
plotly>=5.10.0
datasets>=2.0.0
transformers>=4.20.0
einops>=0.4.1
transformer_lens>=1.0.0
sae_lens>=0.1.0 